Add ESM Build of `@effection/inspect-ui` React Components
I'm embeddding the Effection Inspector into the Backstage UI, and so I
need the ability to access the inspector-ui components directly. In
other words:

```ts
import { App as InspectorApp } from '@effection/inspect-ui';
```

However, the "browser" and "module" entry points don't actually point
to anything, because we don't actually compile them as individual
typescript because the inspector ui was only ever shipped as
an independent application up to this point.

This adds a typescript compilation step and also copies all of the CSS
and SVG assets to the `dist-esm` directory so that the components can
be consumed by 3rd party application.