Add useResource hook.
When embedding the inspector into another react app, we need to load a
slice of data that is continuously updated from the server. This data is loaded
from either a `WebSocket` or an `EventSource` and so it takes some
time to become available as the server and client negotiate could fail at
any point. Once this has been negotiated, and only then, is the stream
set up and the subscription can be used to drive the inspector. It
turns out that it is helpful to model this very Effectionate process
directly in the state

This adds a hook callled `useResource()` that can be used to consume a
resource directly from within a React Component. Resources can either
be `pending`, `resolved`, or `rejected` and as a result, the handle
returned by a resource does not let you access the resource directly,
but via a `Maybe`-like interface called `ResourceHandle`.

```tsx
  const handle = useResource(operation, [dep1, dep2]);

  if (handle.type === 'pending') {
    return <p>Loading</p>;
  } else if (handle.type === 'rejected') {
    return <p>{handle.error.toString()}</p>;
  } else {
    return <MyComponent value={handle.value}/>
  }
```

It's interestig to note that `useResource` can be used with _any_
`Operation` type, and _not_ just `Resource`. This is because normal
operations also provide a value asynchronously. The only difference is
that they are completely done and do not have any on-going computation
associated with them. The following is totaly valid:

```tsx
  const handle = useResource(Promise.resolve('hello'), [dep1, dep2]);

  if (handle.type === 'pending') {
    return <p>Loading</p>;
  } else if (handle.type === 'rejected') {
    return <p>{handle.error.toString()}</p>;
  } else {
    return <MyComponent value={handle.value}/>
  }
```