♻️Migrate to Deno for development workflows
For some time, `v3` work has proceeded using
https://github.com/cowboyd/instructional-effection as the base of
operations. It is now time to settle on that implementation for `v3`
and begin migrating that codebase into this one.

This gets rid of all the NPM/yarn based workflows, and imports the tests from
instructional-effection so that we can run them on this repository.