convert `Subscription` from bare operation to iterator.
> fixes https://github.com/thefrontside/effection/issues/693

In order to simplify working with streams and subscriptions, we need
to be able to disambiguate them easily. The relationship between
`Stream` and `Subscription` is the same as that between `Iterable`
-> `Iterator` and `AsyncIterable` -> `AsyncIterator`.

This converts `Subscription` into an API that is highly analogous to
`AsyncIterator`. The `next()` method returns an `Operation` which
yields an `IteratorResult` So in the same way `AsyncIterator` has:

```ts
next(): Promise<IteratorResult>;
```

The `Subscription` api will look like:

```ts
next(): Operation<IteratorResult>;
```

Once we have this in place, it is trivial to create a stream out of
any Subscription:

```ts
function streamable(subscription: Subscription): Stream {
  return ({
    *[Symbol.iterator]() {
      return subscription;
    }
  });
}
```