🔥Evaluate MDX to eliminate build step
We had to have a build step in order to import our MDX modules without
Deno deploy supporting dynamic imports. However, it turns out that MDX
has an `evaluate` function that can be run on trusted sources to
create an MDX module without needing to import it as an ESM module.

This plugs that in. It's not great, but until isolate runtimes support
imports, it's all we've got.