✨Add main() method as top-level entry point.
The `run()` method serves as a bare-bones entry point into any
Effection operation. However, when you are using it to control your
entire applications there are certain things that you always want to
happen.

On all environments:

- print out an error message if an unhandled error occurs
- provide a mechanism to "exit immediately" from the program, and yet
still gracefully perform all shutdown.

On servers:

- initiate graceful shutdown on SIGINT.

On browsers:

- initiate graceful shutdown on the `unload` event.

If you don't do this, you need to write a lot of boiler-plate code to
make sure that you shutdown gracefully whenever possible, and that you
handle errors properly. However, with a single entry point that
contains everything baked in, you get all of it for free. In a word,
`CTRL-C` just works, and on the browser abruptly killing a tab exits
gracefully.

This carries over the "main" package from `v2` with the notable
exception that you can now use an `exit()` operation to immediately
halt the program.

It uses the new main method to implement the web server for the
website, which demonstrates how much clarity it brings.

Other changes of note:

- removed deno.lock file from the top repository because this is a
library and doesn't need a lock file
- added jsdocs. I think we should start doing this from here on out
now that the APIs are becoming stable.
- the main tests require starting a child process, and so you need to
run `deno test -A` to run all the tests.