📄Add documentation for operations
Operations have changed considerably perhaps the most of anything
between v2 -> v3. Instead of being a union of seven different types,
it is now just a single unified interface. We need to document all of
Effection v3, but especially operations.

We're adopting an approach of constantly leaning on the
likenesses and the differences between `async/await` and
operations. This is because in truth, Effection is really a superset
of `async/await` that provides structural concurrency guarantees.

|    Async      |     Effection       |
|---------------|---------------------|
|Promise        | Operation           |
|await          | yield*              |
|async funciton | function*           |
|AsyncIterable  | Stream              |
|`for await`    | `for yield* iterate`|

As such, this tries to constantly highlight while Operations are like
promises but better.