✨Add all()
The docs use `all()` and `race()` extensively, so we need to have at
least some plausible versions of these functions.

This ports over the tests from `v2`, and is implemented as about the
simplest version of `all()` that I could get working. However, there
is a failing test, and while I think that we can ship it while it
still fails, we do need to fix it before releasing `3.0.0`.

It fails because we need to catch an error in a resource while it is
initializing. We could have any number of resources running, and if
any of them fails before the all is completed, we need to be able to
catch it.