🐛main operation was failing without explicit exit.
`main()` only works correctly whenever you call `yield* exit()` or
if there was a error, because it relies on resolving a containing
action (the entire main method is just one big action).

However, if your operation completes successfully and you _don't_
invoke the exit operation, then it fails because the main action never
resolves. This was because we were returning early from the action.

This removes the return statement and places the implicit exit into a
more prominent location so that we are always guaranteed to resolve
ith an exit status no matter what.