♻️ Make each task run in its own frame.
There are a lot of problems with the way execution is currently
handled that are outlined in this issue
https://github.com/thefrontside/effection/issues/728 owing to the
fact that there can be many tasks executing against the same frame.

This removes that capability, and now the `Block` interface is really
private to the `Frame` object. This makes the code a lot simpler.

However, there is one tradeoff that was necessary to make in order for
this to work. You can now no longer start "bare resources" with
`scope.run()` because they will shut down the frame containing them as
soon as the instruction that allocates the resource is
finished. Instead, if you want to run a resource from a scope, you
will have to suspend it.

```js
scope.run(function*() {
  yield* useResource();
  yield* suspend();
});
```