Add index page, and update introduction.
The intro guide served as both quick pitch deck, and a familiarization
with the concepts. Now that our website is just an HTML page, and not
embedded into doc-a-saurus, we can easily have a nice index page.

This pulls out the enumeration of features into an responsive index
page that has the Effection logo and quick sell.

It also adds jsdoc to the `Operation` type.