✨for yield* each loops
One of the design goals of Effection is to provide structured
concurrency primitives that are very aligned with ordinary JavaScript
and let libraries provide the higher abstractions based on those
primitives. This is why we have `Streams` for AsyncIterables, and
`Subscription` for `AsyncIterator`. What was missing was a natural way
to consume streams that felt like the `for await of` loop.

This introduces the `each()` and `each.next` operations that let you
consume any stream completely naturally.

```javascript
for (let value of yield* each(stream)) {
  console.log(value);
  yield* each.next;
}
```