♻️🔥Simplify the Scope API
Step one of https://github.com/thefrontside/effection/issues/751

This rips out all the child tracking which is already the job of a
frame. Also, because there is now a 1:1 relationship between frame and
scope, there is also, by extension, a 1:1 relationship between frame
and task, and also scope and task. We can make the task lazy, and just
use it as our shutdown logic.