✨Add `Signal` and `Queue` interfaces for more stream creation
It turns out that `Channel` wasn't the perfect primitive because it
requires a scope in which to `send()`. This is super annoying for
callbacks because they are from JavasScript -> Effection and you don't
have a scope. Rather than make one up, this introduces a
`createSignal` function which allows you do directly dispatch events
to any number of consuming operations.

This creates a low level "Queue" class which can be used to directly
manage a subscription. Signal is implemented on top of that, and then
channel is implemented directly on top of that.