🐞Fix arguments passed to NodeJS
The process argument handling is different from Deno and
NodeJS. The following script in Deno:

``` shellsession
$ deno script.ts hello world
```

will set `Deno.args` to `["hello", "world"]`. By contrast, the
following in NodeJS:

``` shellsession
$ node script.mjs hello world
```

will set `process.argv` to
`["/path/to/node","/abs/path/to/script.mjs", "hello", "world"]`

The NodeJS way is similar to the argv of a `Bash` script, but the
Deno way is more practical.

This normalizes Node to behave like Deno, only passing the user
arguments to the script. We can explore how to get the script name and
the executable name later.