Don't create a new frame just to wait for subscription.next()
We were experiencing serious performance degradations when delivering
thousands of messages in a short time. The reason was due to incessant
garbage collection. This was because It was creating a "throw away"
`Frame` whose only purpose was to sit there until the next message was
received at which point it was dropped, and the next time we were
creating.

```js
while (true) {
  let next = subscription.next(); // <-- new Frame every time
  ...
}
```

But a `Frame` is a very heavyweight object just to hold up the current
execution until it can be resumed. In cases where you have many
message deliveries at once, you can end up creating and throwing away
thousands of frames in a matter of milliseconds.

With this change, the `Queue` subscription uses a new `pause`
operation which wraps the _current frame_'s continuation (rather than
allocating a completely new one). This means that there is one frame,
and one generator per subscription, not hundreds of frames, and
hundreds of generators per subscription.

The `pause()` operation is left private to inside Effection for now
because it's unclear what form its API will ultimately take, but it
has several key limitations that make it different from an action:

1. it takes a synchronouse function as its callback, not an operation
2. It returns a function that reverses the effects of that function.

All it does is just pause and resume the _current_ iterator from the
_current_ frame.