Bundle v2 documentation as release artifact.
As part of the move to v3, we're going to be moving the current
documentation from `/effection/` -> `/effection/v2/`. The new
website based on Deno deploy will simply serve this site at a sub
url. However, in order to do this, we need to be able to access this
website as a single artifact that can be downloaded and served.

This creates a workflow on the `v2` branch that, when it sees a tag
conforming to `docs-v2-r(N)` It will build the documentation and
publish it as a GitHub Release. That way, the main documentation site
can download it, cache it, and serve it as a static set of files from
the `/v2/` subpath.

The release is comprised of both the website and the apidocs. This
allows us to update v2 documentation and deploy it to the main website
as needed.