Make V3 site mountable on any other site
The actual Effection website application is running at
https://effection.deno.dev, but it will be served to the public
as https://frontside.com/effection. This means that the application
itself needs to be a chameleon and act as though it is being served
from multiple urls.

To accomplish this, we look for a header `X-Base-Url` coming from the
web request. If this is set, we do two things:

1. this is the base url used to create absolute url for things like
`<meta name="twitter:image">`
2. We prepend the base path of this url to all absolute paths that we
find in the HTML document before it is returned. This includes `<a
href>` and `<img src>` tags.

This is similar to the approach that we use with the graphgen
websites, and platformscript websites. However in those cases, we
stop short of re-writing `href` and `src` attributes and instead
convert every absolute path into an absolute url which is equivalent,
but more cumbersome.