Update the v2 site to promote v3
Now that the V3 site is live at https://frontside.com/effection we
need to point to it from various places.

This updates the legacy site links to point to the new v3
places. Also, because this is now the "prod" version of this site, we
don't need to maintain a separate "prod" config.