Upgrade v2 docs
The V2 docs were upgraded to provide a nudge to the v3 website and
make sure you could link from therer.

This updates the baseline version of the V2 docs to `4`, but it also
adds an environment variable "V2_DOCS_REVISION" which allows us to
upgrade via setting an environment variable.