Tag v3 releases as `v3` (not latest)
We're in a game of whack-a-mole, where every time we publish from the
`v3` branch, that npm package becomes `latest`, and whenever we
publish from the `v2` branch, _that_ package becomes latest.

This explicitly sets the tag used to the `v3`, that way it will not
use `latest`.

Correspondingly, the `v2` tag should be set along with the `latest`
tag. Then, once v3.0.0 is released, it will receive both `v3` and
`latest` and `v2` will only receive `v2`