Don't load all docs just to render the sidebar.
In order to render 1 doc, you need to load them all. This is because
the `title` attribute of each guide is stored inside the doc itself
within the YAML frontmatter. Which means in order to render the title
in the side bar, you have to load and parse all of the the documents.

This pops the title into `structure.json`. That way, we can construct
all the doc metadata without ever having to load or parse anything but
the document that is being requested.

This actually simplifies things considerably, as we can now hang the
list of topics on the doc itself which means no need for a separate
`getTopics()` operation, and no need to maintain a document context.