Express `each()` as a function
We were expressing the `each()` operation as a `const`
reference. Honestly, I can't recall why this was done other than some
vague connection to TypeScript. However, it makes the typedocs weird
because `each` is represented as a `var`, not a `function`.

This switches `each()` to be a function and fixes the typedoc at the
same time.