Revolutionize the website
I've extracted the HTML framework that is driving the Effection
website and the frontside.com website into a separate repo. That way
we don't have to copy and paste between the two. It's also this
framework that will, in all probability, drive the v3 inspector.

It's a middleware stack that is divided into three layers: `jsx`, `html`, and
`http`. The JSX stack maps `Request` to JSX elements, then, once you
have your document, you can transform it as a complete HTML document
syntax tree.

This will let us have interactive components, and it will also let us
peform caching at the HTTP layer.