Task.halt() always succeeds as long as teardown succeeds.
There are two distinct ways in which a task can fail. It can straight
up fail if there is an error during its execution, or it can fail after it has
already completed its execution and somewhere in the teardown process
an error occurs. So for example, a task whose execution succeeds but
whose teardown fails will be considered a failure.

It is also possible to have a task whose execution fails, but whose
teardown succeeds. However, we don't represent this when evaluating
the `Task.halt()` operation. Instead, we always fail if the task
execution has failed. This is a not ideal because if a user awaits or
yields to `Task.halt()` they haven't asked for the result of the task
itself, only for the result of _halting_ it. This makes meta-level
apis like wait groups more difficult because you have to have separate
logic for errored tasks and for successful tasks, when really in some
cases, you want to treat them all as something that just needs to be
halted.

Luckily, we are already tracking the success or failure of frame
destruction internally, so this just adds that information to the
`FrameResult`, so now with every frame, you can not only read both how
its execution went (error, success, abort), but also how went its
destruction.