If an uncaught error happens in Scope.run(), crash the scope.
Right now, you can drop errors on the floor if you call `scope.run()`
and the operation fails. The error will just dissappear, which goes
against the foundations of structured concurrency.

Instead, we want to force folks to deal with the error one way or the
other if they don't want it to crash their stack.

This makes `scope.run()` behave more like `spawn()` where if the error
is not caught it crashes its parent.

Because scopes can now crash and become totally invalid, calling
`Scope.run()` when the scope is inactive is now an error.