✨ Add sitemap.xml to website
We're moving to provide sitemaps for all our minisites as a mechanism
to composing them into the main frontside.com website. This adds a
sitemap plugin which provides a `/sitemap.xml` route that scans all
application middlewares  and if any of them produce a set
of sitemap paths, it addes them to the sitemap.

Right now, the sitemap plugin provides an API compatible version of
the `route()` helper from revolution. The idea is that we can copy
this plugin into our sites and use it until we settle on a good api.
Then, we can either add it to revolution itself or maybe extract it
into a contrib-like package.