✨Introduce Do() operation
Many of the Operations are comprised of a single Effect, and very
often a synchronous Effect at that. There is no need to pay any of the
overhead for creating a generator which has, among other things its
own independent stack frame, when we could just make a simple
iterator.

This introduces a `Do()` operation which takes any effect and converts
it automatically into an operation where it can be used anywhere. This
gives us better performance and also solves the circular dependency
problem inside `context.ts`