⬆️ Upgrate to Deno 2
Deno 2 is out, and we're not going back.

Deno 2 formats yaml files, and it uses a stricter mode of TypeScript
which treat the argument to a catch clause as `unknown` rather than
`any`. In general, this is a good thing, but we have been treating
errors as errors thus far with little bad effect. This adds an
explicit cast to each catch clause to assume that it is an `Error`
like before.

It highlights the danger which is good, and we should seriously
consider moving away from this assumption (although a pox upon the
house of anyone who throws anything but an error). Nevertheless, now
is not the time.