📝 update the `suspend()` docs to account for v4 API
The suspend operation is no longer fundamental. It is just an action
that does nothing, so we had to update the docs to reflect that.

We had to change not only the API docs, but I also noticed that there
wer several places where the old `action+suspend` trick was being
used that are no longer valid.