📝 document error boundaries and the `scoped()` function
The way error boundaries are created has changed from `v3` to `v4`.
Instead of using `call()` or `action()` to establish an error
boundary, there is now a dedicated `scoped()` function which not only
traps any errors, but also creates a resource boundary that does not
let any resource or task escape.

This updates the error documentation, and also adds documentation to
the `scoped()` function. Finally, there was an old deno link that I
updated to JSR.