✨ Backport `withResolvers()` from v4 to v3
`withResolvers()` is a key API and for continuing back to a point and
are necessary to fill hole left by deprecating the operation version
of `action()`.

This is pretty much a straight copy and paste. Some of the imports had
to be re-arranged, and the expect library needed to be upgraded.

It also brings the rosetta stone entry forward