🎒backport advanced scope helpers from v4 -> v3
In v4, the `Scope` interface got a little more powerful 💪💪

Specifically, it got an `expect()` helper to throw an error if a
context is missing. It's the scope equivalent of `Context.expect()`.
Also, there is a `hasOwn()` helper which is necessary for some cases
where there needs to be a single context value per task.

Finally, there is `delete()` which is very rarely used, but rounds out
the API, and since it's present in v4, it's probably a good idea to
add it.